Optimize playground folder layouts to promote efficient development and collaboration
