+++ .idea/gradle.xml
@@ -7,10 +7,9 @@
         <compositeConfiguration>
           <compositeBuild compositeDefinitionSource="SCRIPT">
             <builds>
-              <build path="$PROJECT_DIR$/demo-plugin" name="demo-plugin">
+              <build path="$PROJECT_DIR$/playground" name="playground">
                 <projects>
-                  <project path="$PROJECT_DIR$/demo-plugin" />
-                  <project path="$PROJECT_DIR$/demo-plugin/plugin-core" />
+                  <project path="$PROJECT_DIR$/playground" />
                 </projects>
               </build>
               <build path="$PROJECT_DIR$/plugin" name="plugin">
@@ -28,19 +27,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/bom" />
-            <option value="$PROJECT_DIR$/demo" />
-            <option value="$PROJECT_DIR$/demo-1" />
-            <option value="$PROJECT_DIR$/demo-2" />
-            <option value="$PROJECT_DIR$/demo-plugin" />
-            <option value="$PROJECT_DIR$/demo-plugin/plugin-core" />
-            <option value="$PROJECT_DIR$/kmp" />
-            <option value="$PROJECT_DIR$/kmp/android" />
-            <option value="$PROJECT_DIR$/module" />
-            <option value="$PROJECT_DIR$/module/library-1" />
-            <option value="$PROJECT_DIR$/module/library-2" />
-            <option value="$PROJECT_DIR$/module/library-3" />
-            <option value="$PROJECT_DIR$/module/library-4" />
+            <option value="$PROJECT_DIR$/playground" />
             <option value="$PROJECT_DIR$/plugin" />
             <option value="$PROJECT_DIR$/plugin/winds" />
             <option value="$PROJECT_DIR$/plugin/winds-api" />
