+++ playground/demo-plugin/gradle.properties
