+++ playground/demo-plugin/settings.gradle.kts
@@ -1,5 +1,6 @@
 dependencyResolutionManagement {
-  includeBuild("../plugin/")
+  includeBuild("../../plugin/")
+  @Suppress("UnstableApiUsage")
   repositories {
     google()
     mavenCentral()
@@ -8,7 +9,7 @@ dependencyResolutionManagement {
 
   versionCatalogs {
     create("libs") {
-      from(files("${rootDir.parentFile}/gradle/libs.versions.toml"))
+      from(files("${rootDir.parentFile.parentFile}/gradle/libs.versions.toml"))
     }
   }
 }
