+++ playground/kmp/android/src/main/AndroidManifest.xml
