+++ playground/settings.gradle.kts
@@ -1,0 +1,39 @@
+pluginManagement {
+  includeBuild("../plugin/")
+  // error includeBuild("demo-plugin/")
+  repositories {
+    google()
+    mavenCentral()
+    gradlePluginPortal()
+  }
+}
+@Suppress("UnstableApiUsage")
+dependencyResolutionManagement {
+  repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+  repositories {
+    google()
+    mavenCentral()
+  }
+
+  versionCatalogs {
+    create("libs") {
+      from(files("${rootDir.parentFile}/gradle/libs.versions.toml"))
+    }
+  }
+}
+
+// todo rename to Examples
+//  rootProject.name = "Examples"
+
+include(":bom")
+
+include(":demo")
+include(":demo-1")
+include(":demo-2")
+
+include(":kmp:android")
+
+include(":module:library-1")
+include(":module:library-2")
+include(":module:library-3")
+include(":module:library-4")
