+++ settings.gradle.kts
@@ -1,32 +1,20 @@
 pluginManagement {
-    includeBuild("plugin/")
-    includeBuild("demo-plugin/")
-    repositories {
-        google()
-        mavenCentral()
-        gradlePluginPortal()
-    }
+  includeBuild("plugin/")
+  repositories {
+    google()
+    mavenCentral()
+    gradlePluginPortal()
+  }
 }
 @Suppress("UnstableApiUsage")
 dependencyResolutionManagement {
-    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
-    repositories {
-        google()
-        mavenCentral()
-    }
+  repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+  repositories {
+    google()
+    mavenCentral()
+  }
 }
 
 rootProject.name = "Winds"
 
-include(":bom")
-
-include(":demo")
-include(":demo-1")
-include(":demo-2")
-
-include(":kmp:android")
-
-include(":module:library-1")
-include(":module:library-2")
-include(":module:library-3")
-include(":module:library-4")
+includeBuild("playground")
